+++ Pages/Delivery/DeliveryProcess.cshtml
@@ -1,32 +1,36 @@
 ﻿@page
-@model SuspendedOrders
+@model DeliveryProcess
 @{
     ViewBag.Title = "orders";
 }
 
 <div id="modals">
-    <partial name="~/Pages/Shared/PartialViews/OrderModals/_WaitingModal.cshtml" />
+    <partial name="~/Pages/Shared/PartialViews/OrderModals/_AcceptModal.cshtml" />
     <partial name="~/Pages/Shared/PartialViews/OrderModals/_CancelModal.cshtml" />
-
+    <partial name="~/Pages/Shared/PartialViews/OrderModals/_SuspendModal.cshtml" />
 </div>
 
 
 <div class="my-3 container-fluid d-flex justify-content-center">
     <div>
-        <h3>Askıya Alınan Siparişler</h3>
+        <h3>Bekleyen Siparişler</h3>
     </div>
 </div>
 <div class="orders-content">
     <partial name="~/Pages/Shared/PartialViews/_OrderFilterSectionPartialView.cshtml"/>
 
     <div class="action-section">
-        <button id="waitingModalToggleButton" class="btn btn-sm btn-primary rounded-pill ripple-fx no-border">
+        <button id="acceptModalToggleButton" class="btn btn-sm btn-primary rounded-pill ripple-fx no-border">
             <span class="btn-icon fas fa-bookmark"></span>
-            <span class="btn-title">Bekleme Durumuna Al</span>
+            <span class="btn-title">Sipariş Kabul</span>
         </button>
         <button id="cancelModalToggleButton" class="btn btn-sm btn-danger rounded-pill ripple-fx no-border">
             <span class="btn-icon fas fa-ban"></span>
             <span class="btn-title">Sipariş İptal</span>
+        </button>
+        <button id="suspendModalToggleButton" class="btn btn-sm btn-secondary rounded-pill ripple-fx no-border">
+            <span class="btn-icon fas fa-hourglass"></span>
+            <span class="btn-title">Askıya Al</span>
         </button>
         <button id="refreshGridButton" class="btn btn-sm btn-success rounded-pill ripple-fx no-border">
             <span class="btn-icon fas fa-sync"></span>
@@ -47,4 +51,4 @@
 
 <script src="~/js/dataGrid.js" asp-append-version="true"></script>
 <script src="~/js/orders/orders.js" asp-append-version="true"></script>
-<script src="~/js/orders/suspendedOrders.js" asp-append-version="true"></script>
+<script src="~/js/orders/waitingOrders.js" asp-append-version="true"></script>
