+++ Pages/Orders/WaitingOrders.cshtml
@@ -49,5 +49,6 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/exceljs/3.3.1/exceljs.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/2.0.2/FileSaver.min.js"></script>
 
+<script src="~/js/dataGrid.js" asp-append-version="true"></script>
 <script src="~/js/orders/orders.js" asp-append-version="true"></script>
 <script src="~/js/orders/waitingOrders.js" asp-append-version="true"></script>
