+++ wwwroot/js/orders/orders.js
@@ -16,72 +16,20 @@ const endDate = $("#endDate").dxDateBox("instance");
 const orderNo = $("#orderNo");
 const sku = $("#sku");
 
-const gridContainer = $("#gridContainer");
-
-function formatDate(mydate) {
-    if (mydate === null)
-        return null;
-
-    try {
-        var dateObj = new Date(mydate);
-
-        var monthNames = ["Ocak", "Şubat", "Mart", "Nisan", "Mayıs", "Haziran", "Temmuz", "Ağustos", "Eylül", "Ekim", "Kasım", "Aralık"];
-        var curr_date = dateObj.getDate();
-        var curr_month = dateObj.getMonth();
-        curr_month = curr_month + 1;
-        var curr_year = dateObj.getFullYear();
-        var curr_min = dateObj.getMinutes();
-        var curr_hr = dateObj.getHours();
-        var curr_sc = dateObj.getSeconds();
-        if (curr_month.toString().length === 1)
-            curr_month = '0' + curr_month;
-        if (curr_date.toString().length === 1)
-            curr_date = '0' + curr_date;
-        if (curr_hr.toString().length === 1)
-            curr_hr = '0' + curr_hr;
-        if (curr_min.toString().length === 1)
-            curr_min = '0' + curr_min;
-
-        return curr_year + "-" + curr_month + "-" + curr_date + "T " + curr_hr + ":" + curr_min + ":" + curr_sc + ".0000000Z";
-    } catch (e) {
-        console.log(e);
-        return '';
-    }
-}
-
-const toggleModal = id => {
-    let array = gridContainer.dxDataGrid("instance").getSelectedRowsData();
-    if (array.length === 0) return;
-
-    var s = "";
 
 
-    for (let i = 0; i < array.length; i++) {
-        if (i != array.length - 1)
-            s += array[i].orderno + ", ";
-        else
-            s += array[i].orderno + " ";
-    }
+class OrderGrid extends DataGrid {
 
-    let plu = array.length > 1 ? 'ler' : '';
+    constructor(state, columns) {
+        super("https://tedarikportalorder.azurewebsites.net/api/v1/orders/",
+        'getOrders', "pkey", columns);
 
-    $(id).find(".container").html(s + `sipariş numaralı sipariş${plu} seçildi.`);
-    $(id).modal("show");
-}
-
-class OrderGrid{
-
-    constructor(baseUrl, getMethod, state, columns){
-        this.baseUrl = baseUrl;
-        this.getMethod = getMethod;
         this.state = state;
-        this.columns = columns;
-
-        
-        this.getOrdersAndUpdateTable();
     }
 
-    async getOrdersAndUpdateTable (){
+    
+
+    async getUpdateArray(){
         let bDate = null;
         if (beggingDate)
             bDate = beggingDate.option("value");
@@ -91,7 +39,7 @@ class OrderGrid{
             eDate = endDate.option("value");
 
 
-        let request = {
+        let updateRequest =         {
             "vendor": sessionStorage[vendorKey],
             "state": this.state,
             "orderdatefrom": formatDate(bDate),
@@ -101,100 +49,18 @@ class OrderGrid{
         };
 
 
-        let data = await fetchData(this.baseUrl + this.getMethod, request);
+        let data = await fetchData(this.baseUrl + this.getMethod, updateRequest);
 
         for (let i = 0; i < data.orders; i++) {
             data.orders[i].orderdate = new Date(data.orders[i].orderdate).formatDate("yyyy-MM-dd").toString();
         }
 
-
-        gridContainer.dxDataGrid({
-            dataSource: data.orders,
-            keyExpr: "pkey",
-            columns: this.columns,
-            showBorders: true,
-            noDataText: "Kayıt Bulunamadı",
-            allowColumnResizing: true,
-            rowAlternationEnabled: true,
-            columnAutoWidth: true,
-            filterRow: {
-                visible: true,
-                applyFilter: "auto"
-            },
-            searchPanel: {
-                visible: true,
-                width: 240,
-                placeholder: "Ara..."
-            },
-            scrolling: {
-                mode: "standart"
-            },
-            paging: {
-                enabled: true
-            },
-            headerFilter: {
-                visible: true
-            },
-            selection: {
-                mode: 'multiple'
-            },
-            columnChooser: {
-                enabled: true,
-                mode: "select",
-                title: "Kolon Seçimi"
-            },
-            groupPanel: {
-                visible: true,
-                emptyPanelText: "Gruplamak için buraya sürükleyin"
-            },
-            export: {
-                enabled: true,
-                allowExportSelectedData: true,
-                texts: {
-                    exportAll: "Tümü",
-                    exportSelectedRows: "Seçimi Aktar",
-                    exportTo: "excel"
-                },
-            },
-            onExporting: function (e) {
-                var workbook = new ExcelJS.Workbook();
-                var worksheet = workbook.addWorksheet('Data');
-
-                DevExpress.excelExporter.exportDataGrid({
-                    component: e.component,
-                    worksheet: worksheet,
-                    autoFilterEnabled: true
-                }).then(function () {
-                    workbook.xlsx.writeBuffer().then(function (buffer) {
-                        saveAs(new Blob([buffer], { type: 'application/octet-stream' }), 'Data.xlsx');
-                    });
-                });
-                e.cancel = true;
-            }
-
-        });
-
-        gridContainer.dxDataGrid("columnOption", "pkey", "visible", false);
-
+        return data.orders;
     }
 
-    getSelectedKeys(){
-        let gridInstance = gridContainer.dxDataGrid("instance");
-        return gridInstance.getSelectedRowKeys();
-    }
 
-    async sendData (button, method, reason = null, isCancel = false){
-        let array = this.getSelectedKeys();
-        if (array.length === 0) return;
-
-        let request = {
-            orderKeys: array,
-            isCancel: isCancel,
-            username: "string",
-            lang: "string"
-        };
-
-        if (reason) request["reason"] = reason;
+    async sendData(button, method, reason = null, isCancel = false) {
+        if (this.selectedKeys.length === 0) return;
 
 
         let text = button.html();
@@ -207,16 +73,27 @@ class OrderGrid{
         `
         );
 
-        let data = await fetchData(baseUrl + method, request, false);
+
+        let request = {
+            orderKeys: this.selectedKeys,
+            isCancel: isCancel,
+            username: sessionStorage[vendorNameKey],
+            lang: "string"
+        };
+
+
+        if (reason) request["reason"] = reason;
+
+
+        let data = await fetchData(this.baseUrl + method, request, false);
 
         button.html(text);
 
 
         if (data) {
             button.parents().find("div.modal").modal("hide");
-            this.getOrdersAndUpdateTable();
-        }
-        else {
+            this.updateGrid();
+        } else {
             button.parents("div.modal:first").find(".container").html(
                 /*html*/
                 `<span class="text-danger"><b>Islem Basarisiz</b></span>`
@@ -224,17 +101,25 @@ class OrderGrid{
         }
     }
 
-    async refreshButtonAction(button) {
-        let text = button.html();
 
-        button.html(
-            /*html*/
-            `         
-            <div class="spinner-border text-white" role="status">
-            <span class= "sr-only"> Loading...</span>
-            </div>
-        `
-        );
-        this.getOrdersAndUpdateTable().then(() => button.html(text));
+    showModal(id) {
+        let array = this.gridContainer.dxDataGrid("instance").getSelectedRowsData();
+        if (array.length === 0) return;
+
+
+        var s = "";
+
+        for (let i = 0; i < array.length; i++) {
+            if (i != array.length - 1)
+                s += array[i].orderno + ", ";
+            else
+                s += array[i].orderno + " ";
+        }
+
+
+        let plu = array.length > 1 ? 'ler' : '';
+
+        $(id).find(".container").html(s + `sipariş numaralı sipariş${plu} seçildi.`);
+        $(id).modal("show");
     }
 }
