+++ wwwroot/js/orders/waitingOrders.js
@@ -1,6 +1,4 @@
-const baseUrl = "https://tedarikportalorder.azurewebsites.net/api/v1/orders/";
-
-const orderGrid = new OrderGrid(baseUrl, 'getOrders', "WAI",  [
+const orderGrid = new OrderGrid("WAI",  [
     { dataField: "vendor", caption: "Tedarikçi" },
     { dataField: "orderno", caption: "Siparis No" },
     { dataField: "orderlineno", caption: "Siparis Sira No" },
@@ -25,6 +23,9 @@ searchButton.on("click", () => orderGrid.refreshButtonAction(searchButton));
 refreshGridButton.on("click", () => orderGrid.refreshButtonAction(refreshGridButton));
 
 
-$("#acceptModalToggleButton").on("click", () => { toggleModal("#acceptModal") });
-$("#cancelModalToggleButton").on("click", () => { toggleModal("#cancelModal") });
-$("#suspendModalToggleButton").on("click", () => { toggleModal("#suspendModal") });
+$("#acceptModalToggleButton").on("click", () => { orderGrid.showModal("#acceptModal") });
+$("#cancelModalToggleButton").on("click", () => { orderGrid.showModal("#cancelModal") });
+$("#suspendModalToggleButton").on("click", () => { orderGrid.showModal("#suspendModal") });
+
+
+orderGrid.updateGrid();
